+++ include/lang/GC.h
@@ -155,7 +155,7 @@ public:
 	void Free(PyrObjectHdr* inObj);
 
 
-	int32 StackDepth() { return mVMGlobals->sp - mStack->slots + 1; }
+	long StackDepth() { return mVMGlobals->sp - mStack->slots + 1; }
 	PyrObject* Stack() { return mStack; }
 	void SetStack(PyrObject* inStack) { mStack = inStack; }
 
@@ -343,10 +343,10 @@ inline PyrObject * PyrGC::Allocate(size_t inNumBytes, int32 sizeclass, bool inCo
 	} else {
 		if (sizeclass > kMaxPoolSet) {
 			SweepBigObjects();
-			int32 allocSize = sizeof(PyrObjectHdr) + (sizeof(PyrSlot) << sizeclass);
+			size_t allocSize = sizeof(PyrObjectHdr) + (sizeof(PyrSlot) << sizeclass);
 			obj = (PyrObject*)mPool->Alloc(allocSize);
 		} else {
-			int32 allocSize = sizeof(PyrObjectHdr) + (sizeof(PyrSlot) << sizeclass);
+			size_t allocSize = sizeof(PyrObjectHdr) + (sizeof(PyrSlot) << sizeclass);
 			obj = (PyrObject*)mNewPool.Alloc(allocSize);
 		}
 		if (!obj)
